chore(deps): retry Next.js 16.3.0 and extract shared app shell
Phase 1: bump next and @next/eslint-plugin-next to 16.3.0 (local lint,
test, build pass — Vercel API routing must be verified on preview).

Prep for public App cutover:
- Move authOptions to src/server/auth/authOptions.ts
- Add AppShellProviders shared by Pages _app and AppRootLayoutClient
