+++ package.json
@@ -87,7 +87,7 @@
     "graphql": "16.14.0",
     "joi": "^18.2.3",
     "monaco-editor": "^0.56.0",
-    "next": "16.2.12",
+    "next": "16.3.0",
     "next-auth": "4.24.15",
     "notistack": "^3.0.2",
     "overlayscrollbars": "^2.14.0",
@@ -119,7 +119,7 @@
     "@graphql-codegen/typescript": "^4.1.6",
     "@graphql-codegen/typescript-operations": "^4.6.1",
     "@graphql-codegen/typescript-react-apollo": "^4.4.1",
-    "@next/eslint-plugin-next": "16.2.12",
+    "@next/eslint-plugin-next": "16.3.0",
     "@semantic-release/changelog": "^6.0.3",
     "@semantic-release/git": "^10.0.1",
     "@testing-library/dom": "^10.4.1",
