+++ src/app/AppRootLayoutClient.tsx
@@ -1,27 +1,19 @@
 "use client";
 
-import { ApolloProvider } from "@apollo/client";
 import { AppRouterCacheProvider } from "@mui/material-nextjs/v16-appRouter";
 import { Analytics } from "@vercel/analytics/react";
 import { SpeedInsights } from "@vercel/speed-insights/next";
 import type { Session } from "next-auth";
-import { SessionProvider } from "next-auth/react";
-import { SnackbarProvider } from "notistack";
 import React, { type ReactNode } from "react";
 import "symbol-observable";
 
 import { CookieConsentRoot } from "#/features/cookieConsent/ui/CookieConsentRoot";
 import { ProjectBrowser } from "#/features/project/ui/ProjectBrowser/ProjectBrowser";
 import { ProjectBrowserProvider } from "#/features/project/ui/ProjectBrowser/ProjectBrowserContext";
-import { apolloClient } from "#/graphql/apolloClient";
 import { type I18nProps } from "#/i18n/getI18nProps";
 import type { Locales } from "#/i18n/i18n-types";
-import { TrpcProvider } from "#/shared/trpc/TrpcProvider";
-import { SnackbarCloseButton } from "#/shared/ui/atoms/SnackbarCloseButton";
+import { AppShellProviders } from "#/shared/ui/providers/AppShellProviders";
 import { AppRouterI18nProvider } from "#/shared/ui/providers/I18nProvider";
-import { StateThemeProvider } from "#/shared/ui/providers/StateThemeProvider";
-import { isSnackbarClosable } from "#/shared/ui/snackbarClosability";
-import { ReduxProvider } from "#/store/provider";
 
 type AppRootLayoutClientProps = {
   children: ReactNode;
@@ -31,7 +23,7 @@ type AppRootLayoutClientProps = {
 };
 
 /**
- * Mirrors `pages/_app` providers for App Router routes (default locale home).
+ * App Router provider shell. Shares {@link AppShellProviders} with Pages `_app`.
  */
 export const AppRootLayoutClient: React.FC<AppRootLayoutClientProps> = ({
   children,
@@ -41,43 +33,18 @@ export const AppRootLayoutClient: React.FC<AppRootLayoutClientProps> = ({
 }) => {
   return (
     <AppRouterCacheProvider options={{ key: "css" }}>
-      <TrpcProvider>
-        <ReduxProvider>
-          <SessionProvider session={session}>
-            <ApolloProvider client={apolloClient}>
-              <StateThemeProvider>
-                <SnackbarProvider
-                  maxSnack={4}
-                  action={(snackbarKey) =>
-                    isSnackbarClosable(snackbarKey) ? (
-                      <SnackbarCloseButton snackbarKey={snackbarKey} />
-                    ) : null
-                  }
-                  classes={{
-                    containerAnchorOriginBottomLeft:
-                      "snackbar-mobile-bottom-margin",
-                    containerAnchorOriginBottomCenter:
-                      "snackbar-mobile-bottom-margin",
-                    containerAnchorOriginBottomRight:
-                      "snackbar-mobile-bottom-margin",
-                  }}
-                >
-                  <AppRouterI18nProvider locale={locale} i18n={i18n}>
-                    <CookieConsentRoot>
-                      <ProjectBrowserProvider>
-                        {children}
-                        <ProjectBrowser />
-                        <Analytics />
-                        <SpeedInsights />
-                      </ProjectBrowserProvider>
-                    </CookieConsentRoot>
-                  </AppRouterI18nProvider>
-                </SnackbarProvider>
-              </StateThemeProvider>
-            </ApolloProvider>
-          </SessionProvider>
-        </ReduxProvider>
-      </TrpcProvider>
+      <AppShellProviders session={session}>
+        <AppRouterI18nProvider locale={locale} i18n={i18n}>
+          <CookieConsentRoot>
+            <ProjectBrowserProvider>
+              {children}
+              <ProjectBrowser />
+              <Analytics />
+              <SpeedInsights />
+            </ProjectBrowserProvider>
+          </CookieConsentRoot>
+        </AppRouterI18nProvider>
+      </AppShellProviders>
     </AppRouterCacheProvider>
   );
 };
