+++ src/server/auth/get-server-auth-session.ts
@@ -1,7 +1,7 @@
 import { type GetServerSidePropsContext } from "next";
 import { getServerSession } from "next-auth";
 
-import { authOptions } from "#/pages/api/auth/[...nextauth]";
+import { authOptions } from "#/server/auth/authOptions";
 
 /**
  * Wrapper for getServerSession https://next-auth.js.org/configuration/nextjs
